Implemented core functionality of the Audio module, all untested. Only thing not finished is fade.
Added Ruby attributes to all the existing classes/modules types in the RGSS folder. Still need to do the properties and functions. Although IronRuby automatically changes the names to match the convention of the language, I prefer to do it explicitely and not trust it.

Drastically improved the way renderable objects get drawn. Don't know what I was thinking, but no more adding/removing/sorting each frame. Sprites get added to the render targets list when created, and get removed when they are disposed.